// src/static/enums.tsx
export enum ITEMQUALITY {
    common = "common",
    fine = "fine",
    superior = "superior",
    epic = "epic",
    legendary = "legendary",
    apex = "apex",
    radiant = "radiant",
  }
  
  export enum ITEMTYPE {
    mount = "mount",
    furnishing = "furnishing",
    style="style",
    costume = "costume",
    consumable = "consumable",
    pet = "pet",
    memento = "memento",
    upgrade = "upgrade",
    andornment="andornment",
  }